+++ test/slo.test.js
@@ -47,6 +47,16 @@ describe('slo能够正确响应并注销', function() {
     casClientFactory(casClientApp, {
       servicePrefix: clientPath,
       serverPath: casRootPath,
+      hooks: {
+        before: function(req, res, next) {
+          req.start = Date.now();
+          next();
+        },
+        after: function(req, res, next) {
+          expect(req.start).to.not.be.empty;
+          next();
+        }
+      },
       logger: function(req, type) {
         return function() {
         };